Refactored the project so that each function can be deployed separately to different lambdas
