+++ Task 2/backend/Makefile
@@ -1,2 +1,8 @@
 build-ApiFunction:
 	GOOS=linux GOARCH=amd64 CGO_ENABLED=0 go build -o bootstrap ./cmd/main.go
+
+
+#set GOOS=linux                                           
+#set GOARCH=amd64
+#go build -o bootstrap
+#powershell Compress-Archive -Path bootstrap -DestinationPath function.zip -Force
